requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
Django==2.1.4
django-cors-headers==2.4.0
django-sendsms==0.3.1
djangorestframework==3.9.0
gunicorn==19.9.0
idna==2.8
psycopg2==2.7.6.1
pycparser==2.19
PyJWT==1.7.1
PySocks==1.6.8
python-dateutil==2.7.5
python-decouple==3.1
pytz==2018.7
requests==2.21.0
six==1.10.0
urllib3==1.24.1
